added tests, modularized code into files
